360 degree spinning images

+++ Player.hpp
@@ -10,9 +10,9 @@ class Player {
 public:
     size_t x;
     size_t y;
-    size_t theta;
+    float theta;
 
-    Player(size_t x, size_t y, size_t theta = 0);
+    Player(size_t x, size_t y, float theta = 0);
 
     float find_depth(Window2D& window, float phi);
     std::vector<float> find_pov(Window2D& window, float pov_angle);
